# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /nyaya-ai/chat:
    post:
      operationId: nyayaAiChat
      tags: [nyaya-ai]
      summary: Ask Nyaya AI (Your Rights Guide)
      description: >
        Stateless chat with Nyaya Nagri's ONE AI assistant. Gemini-powered
        and retrieval-grounded: specific legal facts come only from a
        pre-approved corpus of PRD §4 act summaries whose public source is
        India Code (indiacode.nic.in) — the model never invents acts or
        sections. Understands safe game context (current zone, completed
        zones, progress, badges, current lesson) passed as stable ids and
        capped strings. Enforces the identical safety contract as the
        persona route: deterministic distress escalation before any AI call
        (English, Hindi, Hinglish, and Gujarati lexicon), PII ingress
        redaction, untrusted history quoting, and the fail-closed helpline
        output gate. Replies follow the child's own language (English /
        Hindi / Hinglish / Gujarati). Educational legal information, never
        professional legal advice; no conversation data is persisted.
        Requires the server-side GEMINI_API_KEY secret — the key never
        reaches the client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NyayaAiChatInput"
      responses:
        "200":
          description: Nyaya AI reply
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarChatReply"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarError"
        "502":
          description: Upstream AI error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarError"
        "503":
          description: GEMINI_API_KEY is not configured on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarError"
  /nyaya-ai/chat-stream:
    post:
      operationId: nyayaAiChatStream
      tags: [nyaya-ai]
      summary: Ask Nyaya AI — low-latency streaming twin (NDJSON)
      description: >
        Streaming variant of /nyaya-ai/chat with the IDENTICAL deterministic
        safety contract (shared preparation code server-side). The response
        body is NDJSON — one JSON event per line:
        {"type":"delta","text":"..."} is a gated increment of the reply (the
        fail-closed helpline output gate runs over the ACCUMULATED reply
        BEFORE each increment is forwarded, so no ungated text ever reaches
        the client); {"type":"escalated","reply":"..."} carries the canonical
        hard-coded escalation text and the client replaces the whole partial
        reply with it; {"type":"done"} ends a clean reply; {"type":"error"}
        means the upstream died mid-reply (the client keeps the partial text
        and offers a retry). Failures before the first event use plain HTTP
        statuses. Consumed by a small hand-written streaming reader in the
        web client — the generated JSON client cannot parse NDJSON.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NyayaAiChatInput"
      responses:
        "200":
          description: NDJSON event stream (application/x-ndjson)
          content:
            application/x-ndjson:
              schema:
                type: string
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarError"
        "502":
          description: Upstream AI error before any output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarError"
        "503":
          description: GEMINI_API_KEY is not configured on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarError"
...
components:
  schemas:
...
    NyayaAiChatInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 500
        ageBand:
          type: string
          enum: ["8-11", "12-15", "16-18"]
          description: >
            Optional — Nyaya AI is reachable from the Home screen before
            onboarding, where no age band exists yet. When present it only
            tunes tone; it never changes the allowed facts.
        language:
          type: string
          enum: [en, hi]
          description: >
            App UI language (defaults to en). Replies follow the CHILD's own
            message language (English / Hindi / Hinglish / Gujarati); this
            value is the fallback for mixed/unclear input and for canonical
            safety text. Safety behaviour is identical in every language;
            helpline digits never change.
        history:
          type: array
          maxItems: 12
          description: Recent turns kept client-side only (never persisted)
          items:
            $ref: "#/components/schemas/AvatarChatTurn"
        gameContext:
          $ref: "#/components/schemas/NyayaAiGameContext"
      required:
        - message
    NyayaAiGameContext:
      type: object
      description: >
        Safe, app-generated game state for personalization (PRD §9 data
        minimization: stable zone IDS — mapped to names server-side — plus
        capped counts and the fun game nickname; never any real-world PII).
      properties:
        nickname:
          type: string
          maxLength: 24
          description: Fun made-up game nickname (never a real name)
        currentZoneId:
          type: string
          maxLength: 16
          description: Zone the player is inside (e.g. zone3)
        nearbyZoneId:
          type: string
          maxLength: 16
          description: Nearest zone on the map when not inside one
        completedZoneIds:
          type: array
          maxItems: 8
          items:
            type: string
            maxLength: 16
        progressPct:
          type: number
          minimum: 0
          maximum: 100
        badgeCount:
          type: number
          minimum: 0
          maximum: 99
        currentLessonTitle:
          type: string
          maxLength: 120
        currentLevelNumber:
          type: number
          minimum: 1
          maximum: 20
        learnQuestionsAnswered:
          type: number
          minimum: 0
          maximum: 100000
          description: >
            Total recorded quiz/recap answers (learning-insights log). Lets
            Nyaya AI answer "How am I doing?" from real stats only.
        learnAccuracyPct:
          type: number
          minimum: 0
          maximum: 100
          description: >
            Overall recorded-answer accuracy — the client sends it only
            when the analyzer's minimum-evidence gate passed.
        learnTrend:
          type: string
          enum: [improving, steady, declining]
          description: Session-over-session accuracy direction (evidence-gated)
        strongZoneId:
          type: string
          maxLength: 16
          description: Zone with the strongest evidenced accuracy label
        practiceZoneId:
          type: string
          maxLength: 16
          description: Zone the deterministic analyzer suggests practicing
...
    AvatarChatTurn:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          maxLength: 2000
      required:
        - role
        - content
    AvatarChatReply:
      type: object
      properties:
        reply:
          type: string
        escalated:
          type: boolean
          description: True when the safety escalation path produced the reply
      required:
        - reply
        - escalated
    AvatarError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...