# lib/api-zod/src/generated/types/nyayaAiGameContext.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NyayaAiGameContextLearnTrend } from './nyayaAiGameContextLearnTrend';

/**
 * Safe, app-generated game state for personalization (PRD §9 data minimization: stable zone IDS — mapped to names server-side — plus capped counts and the fun game nickname; never any real-world PII).
 */
export interface NyayaAiGameContext {
  /**
     * Fun made-up game nickname (never a real name)
     * @maxLength 24
     */
  nickname?: string;
  /**
     * Zone the player is inside (e.g. zone3)
     * @maxLength 16
     */
  currentZoneId?: string;
  /**
     * Nearest zone on the map when not inside one
     * @maxLength 16
     */
  nearbyZoneId?: string;
  /**
     * @maxItems 8
     * @items.maxLength 16
     */
  completedZoneIds?: string[];
  /**
     * @minimum 0
     * @maximum 100
     */
  progressPct?: number;
  /**
     * @minimum 0
     * @maximum 99
     */
  badgeCount?: number;
  /** @maxLength 120 */
  currentLessonTitle?: string;
  /**
     * @minimum 1
     * @maximum 20
     */
  currentLevelNumber?: number;
  /**
     * Total recorded quiz/recap answers (learning-insights log). Lets Nyaya AI answer "How am I doing?" from real stats only.
     * @minimum 0
     * @maximum 100000
     */
  learnQuestionsAnswered?: number;
  /**
     * Overall recorded-answer accuracy — the client sends it only when the analyzer's minimum-evidence gate passed.
     * @minimum 0
     * @maximum 100
     */
  learnAccuracyPct?: number;
  /** Session-over-session accuracy direction (evidence-gated) */
  learnTrend?: NyayaAiGameContextLearnTrend;
  /**
     * Zone with the strongest evidenced accuracy label
     * @maxLength 16
     */
  strongZoneId?: string;
  /**
     * Zone the deterministic analyzer suggests practicing
     * @maxLength 16
     */
  practiceZoneId?: string;
}